fix: Fixes drag and drop jumping (close #49)
